refactor: Extract VerificationDots to shared component and add German translations
Address Copilot PR review comments:

1. **DRY violation fixed**: Extracted duplicated VerificationDots component
   - Created shared src/components/VerificationDots.tsx
   - Removed duplicate implementations from ActivityDetailDialog and ActivityLogList
   - Added size prop ('sm' | 'md') to control dot dimensions
   - ActivityLogList uses size="sm" (w-2 h-2), DetailDialog uses default "md" (w-3 h-3)

2. **useEffect optimization**: Fixed dependency array in ActivityDetailDialog
   - Changed from entire 'activity' object to specific properties
   - Now uses [activity.id, activity.verification, open]
   - Prevents unnecessary re-renders when unrelated activity properties change

3. **Complete German translations**: Added all missing translations
   - Hash-Kette (Daten) / Hash-Kette (Link)
   - Merkle-Baum
   - OpenTimestamp: Nicht zutreffend
   - {label}: Gültig/Ungültig/Ausstehend/Nicht zutreffend
   - Compiled with npm run lingui:compile
