+++ src/locales/de/messages.mjs
@@ -1,1 +1,1 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"+MTt8f\":[\"\\\"\",[\"unitName\"],\"\\\" löschen?\"],\"+SxetR\":[\"Merkle Tree\"],\"+b7T3G\":[\"Aktualisiert\"],\"+hZPci\":[\"Datum des Vertragsbeginns\"],\"+jEDxY\":[\"Kunde bearbeiten\"],\"+zy2Nq\":[\"Typ\"],\"/4TFrF\":[\"Vollständiger Name\"],\"/AsMcZ\":[\"Aktiviere ...\"],\"/Q5FOE\":[\"Mitarbeiter bearbeiten\"],\"/YtZc9\":[\"Informationen zur Tätigkeit\"],\"/hW+7s\":[\"Willkommen zum SecPal Onboarding\"],\"036J+m\":[\"Zertifikat:\"],\"0I3xLT\":[\"✓ Benachrichtigungen sind aktiviert. Sie erhalten Updates basierend auf Ihren Einstellungen.\"],\"0VfT1o\":[\"Geheimnis laden...\"],\"0qqeZs\":[\"Um einem Mitarbeiter den Dienstgrad zu entziehen, muss der Benutzer die Berechtigung haben, diesen Dienstgrad zu vergeben. Dies verhindert eine Eskalation der Privilegien.\"],\"0tttqe\":[\"Gültig bis\"],\"0wJVK+\":[\"Grundlegend\"],\"18L2W6\":[\"Mitarbeiter anlegen\"],\"18zG7E\":[\"Verwalten Sie Ihre internen Organisationseinheiten wie Abteilungen, Niederlassungen und Teams.\"],\"1A3EXy\":[\"Ausklappen\"],\"1DBGsz\":[\"Anmerkungen\"],\"1GixfT\":[\"Bald ablaufend\"],\"1eiedP\":[\"Organisatorische Struktur\"],\"1ep1s+\":[\"Zuletzt geändert: \",[\"0\"]],\"1k5CRu\":[\"Unterstützt: Bilder, PDFs, Dokumente (max. 10MB)\"],\"1oL0IJ\":[\"Sind Sie sicher, dass Sie diesen Anhang löschen möchten?\"],\"2BBAbc\":[\"Liste\"],\"2GT3Hf\":[\"Global\"],\"2wgink\":[\"Anzeigen von <0>\",[\"0\"],\"</0> bis <1>\",[\"1\"],\"</1> von <2>\",[\"2\"],\"</2> Mitarbeitern\"],\"3T71mA\":[\"Wählen Sie, welche Benachrichtigungen Sie erhalten möchten\"],\"3TgTKY\":[\"Datei konnte nicht gelöscht werden\"],\"3aPOvd\":[\"Anhang konnte nicht gelöscht werden\"],\"3dwitg\":[\"Mitteilungen von Teammitgliedern und Vorgesetzten\"],\"3e8eQb\":[\"Keine organisatorischen Einheiten\"],\"3wV73y\":[\"Stadt\"],\"43Ygvu\":[\"Fehler beim Laden des Geheimnisses\"],\"48oQna\":[\"Das Anlegen von Organisationseinheiten ist im Offline-Modus nicht möglich. Bitte stellen Sie die Verbindung wieder her, um Änderungen vorzunehmen.\"],\"4AY++3\":[\"Standort-Nummer:\"],\"4DZJcB\":[\"Die Bearbeitung von Organisationseinheiten ist im Offline-Modus nicht möglich. Bitte stellen Sie die Verbindung wieder her, um Änderungen vorzunehmen.\"],\"4DjwT6\":[\"Legen Sie fest, welchen Rangbereich dieser Benutzer sehen kann. Lassen Sie das Feld leer oder setzen Sie es auf 0, um nur Mitarbeiter ohne Führungsaufgaben (Sicherheitsmitarbeiter) zu sehen.\"],\"4FcOp+\":[\"Fehler beim Laden eines Mitarbeiters\"],\"4OaTD4\":[\"in Freistellung\"],\"4Z4JSO\":[\"Name des Kunden\"],\"4cxQxp\":[\"Anzeigen von <0>\",[\"0\"],\"</0> bis <1>\",[\"1\"],\"</1> von <2>\",[\"2\"],\"</2> Mitarbeitern\"],\"4fjnRg\":[\"-- Wählen Sie ein Geheimnis --\"],\"4geA9Q\":[\"Fehler beim Laden von Geheimnissen\"],\"4yI3Vk\":[\"Zurück zu Geheimnisse\"],\"4yv9An\":[\"Erweitert\"],\"5/FzWg\":[\"Gelöscht \",[\"count\"],\" abgeschlossene(r) Upload(s)\"],\"58X1WA\":[\"Alle Einheiten\"],\"5A0Vat\":[\"Untergeordnete Einheit hinzufügen\"],\"5HYWB0\":[\"Bereich\"],\"5I+k35\":[\"Mitarbeiter Nummer\"],\"5NZpX8\":[\"Listenansicht\"],\"5ZulCG\":[\"Typ ist erforderlich\"],\"5cVSqz\":[\"Gemeinsam genutzt mit (\",[\"0\"],\")\"],\"5dJK4M\":[\"Rollen\"],\"5eHyY+\":[\"Test-Benachrichtigung\"],\"5fap5A\":[\"Standort nicht gefunden\"],\"5gqNQl\":[\"Gitternetz-Ansicht\"],\"5lq7gm\":[\"Standardmäßig (nicht aktiviert) können Benutzer ihre eigen Personalakte weder sehen noch bearbeiten. Dies verhindert die Selbstmanipulation von Gehalt, Rang usw. Aktivieren Sie dieses Kontrollkästchen nur, wenn Sie möchten, dass dieser Benutzer auf seine eigene Personalakte zugreifen kann.\"],\"5tmzAO\":[\"Die neue Einheit wird unter dieser übergeordneten Einheit erstellt.\"],\"6GBt0m\":[\"Metadaten\"],\"6LhVHS\":[\"Übergeordnete Einheit\"],\"6Q2qjr\":[\"Kunden suchen...\"],\"6VfR2o\":[\"Qualifikationen\"],\"6YtxFj\":[\"Name\"],\"6gRgw8\":[\"Wiederholen\"],\"6kPk3+\":[\"Persönliche Informationen\"],\"6splsl\":[\"Ungültige Kombination: Es werden keine Mitarbeiter mit diesem Bereich angezeigt.\"],\"6yvfgr\":[\"Benachrichtigungen aktiviert\"],\"7/EICV\":[\"- Schreiben: Ansehen + Bearbeiten Geheimnis\"],\"7/tqYA\":[\"neue übergeordnete Einheit auswählen\"],\"7Att6/\":[\"Qualifikationen laden...\"],\"7L01XJ\":[\"Aktionen\"],\"7OWvhx\":[\"OpenTimestamp Bestätigt\"],\"7TJx4N\":[\"Eine neue Version von SecPal ist verfügbar.\"],\"7i8j3G\":[\"Unternehmen\"],\"7sNhEz\":[\"Benutzername\"],\"7uHWR+\":[\"Geburtsdatum\"],\"8/brI5\":[\"Name ist erforderlich\"],\"842lnw\":[\"Erstellen Sie Ihr erstes Geheimnis, um loszulegen\"],\"84T3OZ\":[\"Maximal zuordenbarer Rang\"],\"876pfE\":[\"Kunde\"],\"89yiwg\":[\"Der Benutzer kann KEINEN Rang zuweisen ODER entfernen (verhindert eine Eskalation der Rechte).\"],\"8CnLST\":[\"Geheimnisse ansehen\"],\"8RYNR1\":[\"Daten aktualisieren\"],\"8Tg/JR\":[\"Benutzerdefiniert\"],\"8ZsakT\":[\"Passwort\"],\"8bl30j\":[\"Aktualisieren Sie die Details der Organisationseinheit.\"],\"8rE61T\":[\"Rechnungsadresse\"],\"8vETh9\":[\"Anzeigen\"],\"8z4jWY\":[\"Sicherheitswarnungen\"],\"9AcM8x\":[\"Löschen Abgeschlossen\"],\"9H0JwN\":[\"Mitarbeiter #\"],\"9P/b/5\":[\"App-Updates und Wartungsbenachrichtigungen\"],\"9bEGEb\":[\"Läuft ab:\"],\"9cDpsw\":[\"Berechtigungen\"],\"9cRAwD\":[\"← Zurück zu Geheimnissen\"],\"A1taO8\":[\"Suche\"],\"A8qnCR\":[\"Dies ist eine Testbenachrichtigung von SecPal\"],\"AAnV/D\":[\"Wählen Sie eine neue übergeordnete Stelle für diese Organisationseinheit aus.\"],\"AN9mgS\":[[\"filename\"],\" aus der Warteschlange entfernt\"],\"AbIhll\":[\"Vertragsart\"],\"AuwJQ1\":[\"Ungültiger Dateityp: \",[\"0\"],\". Nur Bilder, PDFs und Dokumente sind erlaubt.\"],\"BHoh+C\":[\"Minimaler zuweisbarer Rang\"],\"BLXWJv\":[\"Keine Kunden gefunden\"],\"Baxs4G\":[\"Standort löschen\"],\"BfQvQ5\":[\"System-Updates\"],\"Bgjf1M\":[\"Geheimnis teilen fehlgeschlagen\"],\"BgqIX/\":[\"OpenTimestamp: N/A\"],\"BltQR1\":[\"Standorte anzeigen\"],\"C+ffQE\":[\"Vollzeit\"],\"C448X5\":[\"z.B. Niederlassung Berlin\"],\"CCXa60\":[\"Keine Dokumente hochgeladen\"],\"CEwoiy\":[\"Niedrigste Rangnummer, die der Benutzer sehen kann (1 = CEO, höhere Zahlen = niedrigere Positionen). Bereich: 1-255\"],\"CXDHcv\":[\"Raster\"],\"CjXztc\":[\"Zugangskontrolle\"],\"ClD8Z1\":[\"Details zur Beschäftigung\"],\"Co/eKl\":[\"Benutzer können ihre eigenen HR-Daten anzeigen/bearbeiten\"],\"Cyc2u5\":[\"Aktivitätsprotokollinformationen und Überprüfungsstatus anzeigen\"],\"CzsL7L\":[\"Teilzeit\"],\"D+NlUC\":[\"System\"],\"D2aVgL\":[\"Vorheriger Hash\"],\"D5Y6ym\":[\"Schritt 3: Kontrolle des Selbstzugriffs\"],\"DBoLwV\":[\"Alle Tags\"],\"DBr12o\":[\"Anhänge (\",[\"0\"],\")\"],\"DDffEz\":[\"Maximal sichtbarer Rang\"],\"DHhJ7s\":[\"Vorherige\"],\"DYlwPh\":[\"Kontaktperson\"],\"DebAtR\":[[\"errorOps\"],\" Operation(en) fehlgeschlagen\"],\"Du6bPw\":[\"Adresse\"],\"E/muDO\":[\"Dokumente\"],\"EF2EU9\":[\"Löschen...\"],\"EH48dY\":[\"Einige Funktionen sind möglicherweise eingeschränkt. Ihre Änderungen werden synchronisiert, sobald Sie wieder online sind.\"],\"ELI7SU\":[[\"label\"],\": Pending\"],\"ELa85n\":[\"Alle Typen\"],\"EOv8gS\":[\"Koordinaten\"],\"EPp7ex\":[\"Der Name darf maximal 255 Zeichen lang sein\"],\"EkH9pt\":[\"Update\"],\"ErGYe4\":[\"Aktualisieren Sie die Angaben zu Ihrem Geheimnis.\"],\"F6pfE9\":[\"Aktiv\"],\"F7ErMS\":[[\"label\"],\": Invalid\"],\"FEBEYO\":[\"Geheimnisse durchsuchen...\"],\"FQBaXG\":[\"Aktivieren\"],\"FQfoqR\":[\"Ihre lokale Version\"],\"FSimqX\":[\"Kunde auswählen...\"],\"FYzHRF\":[\"Holding\"],\"Failed to change language. Please try again.\":[\"Die Sprache konnte nicht geändert werden. Bitte versuchen Sie es erneut.\"],\"FjhOTo\":[\"Kunde löschen\"],\"Flmwkt\":[\"Mitarbeiter\"],\"FtJ+Sn\":[\"Meine Organisation\"],\"FxAKnZ\":[\"Kritische Sicherheitsbenachrichtigungen und -warnungen\"],\"GW+FsP\":[\"Auswahl anwenden\"],\"GXsAby\":[\"Widerrufen\"],\"GhMjr9\":[\"Dateien verschlüsseln...\"],\"Gwmfl9\":[\"Detailfenster schließen\"],\"GxkJXS\":[\"Hochladen...\"],\"GyKiE+\":[\"Untergeordnete Einheiten müssen gelöscht oder einer anderen übergeordneten Einheit neu zugeordnet werden, bevor diese Organisationseinheit entfernt werden kann.\"],\"GyrqoW\":[\"Aktivitätsprotokolle\"],\"H86f9p\":[\"Zusammenklappen\"],\"H8BDKE\":[\"Zurück zur Liste\"],\"HBk9Vi\":[\"neue übergeordnete Einheit\"],\"HM56Bx\":[\"Erstellen...\"],\"HVUgrA\":[\"Die Bereichszuweisung konnte nicht gespeichert werden.\"],\"IUwGEM\":[\"Änderungen speichern\"],\"IXHBs4\":[\"Standort anlegen\"],\"IagCbF\":[\"URL\"],\"Id+qAX\":[\"Keine Aktivitätsprotokolle gefunden.\"],\"IiFHaM\":[\"Wählen Sie Ihre bevorzugte Sprache für die Anwendung.\"],\"Iv4GiC\":[\"Der Typ bestimmt die Rolle der Einheit in der Hierarchie.\"],\"IvB46R\":[\"Zur Ansicht herunterladen\"],\"J00gME\":[\"Anzeigen von <0>\",[\"0\"],\"</0> bis <1>\",[\"1\"],\"</1> von <2>\",[\"2\"],\"</2> Standorten\"],\"J4cXlI\":[\"OpenTimestamp\"],\"JHB/Ky\":[\"Invalid\"],\"JLJLL1\":[\"Untergeordnete Einheit hinzufügen\"],\"JNCzPW\":[\"Land\"],\"JQEpPE\":[\"Sicherheitsstufe\"],\"JRVmaF\":[\"Hochgeladen: \",[\"0\"]],\"Ju2mNv\":[\"Keine Geheimnisse gefunden\"],\"K0pe4j\":[\"Alle Protokolle\"],\"K7P0jz\":[\"Zuletzt aktualisiert\"],\"K8Qnlj\":[\"Verschiebe...\"],\"KELjcQ\":[\"Willkommen zu SecPal\"],\"Kfa+gn\":[\"Vorvertrag\"],\"KhU0au\":[\"Nur offline\"],\"KnN1Tu\":[\"Läuft ab\"],\"Kzoh+W\":[\"SecPal - der beste Freund eines Wachmanns. Eine Offline-First Progressive Web App für Sicherheitspersonal, die digitale Wachbücher mit modernem Service-Management verbindet.\"],\"LCDse8\":[\"Automatisches Aktualisieren (30s)\"],\"LII+b3\":[\"Wiederholungsversuch geplant. Dateien werden hochgeladen, wenn das Netzwerk verfügbar ist.\"],\"Ln8nFJ\":[\"Ziel teilen\"],\"LoUIHT\":[\"Minijob\"],\"LtI9AS\":[\"Eigentümer\"],\"M1RnFv\":[\"Abgelaufen\"],\"M3sRlm\":[\"Fehler beim Laden von Mitarbeitern\"],\"M5lRtC\":[\"Läuft ab: \",[\"0\"]],\"MAv87s\":[\"Geheimnisse\"],\"MLW/Q/\":[\"Das Verschieben von Organisationseinheiten ist im Offline-Modus nicht möglich. Bitte stellen Sie die Verbindung wieder her, um Änderungen vorzunehmen.\"],\"MpF6Ix\":[\"Upload-Warteschlange\"],\"MvsnLi\":[\"Upload-Wiederholungsversuch konnte nicht geplant werden\"],\"MyWb0G\":[\"Verwalten Sie Ihre Geheimnisse und sensiblen Informationen auf sichere Weise.\"],\"N40H+G\":[\"Alle\"],\"NgNAlL\":[\"Schritt \",[\"currentStep\"],\" von \",[\"totalSteps\"]],\"NihQNk\":[\"Kunden\"],\"NoNwIX\":[\"Inaktiv\"],\"Np2enW\":[\"Wählen Sie eine neue übergeordnete Einheit oder machen Sie diese zu einer Stammeinheit.\"],\"NtEny3\":[\"Verschlüsselungsschlüssel für Anhänge kann nicht geladen werden. Bitte aktualisieren Sie die Seite oder kontaktieren Sie den Support, falls das Problem weiterhin besteht.\"],\"Nu4DdT\":[\"Synchronisieren\"],\"Nu4oKW\":[\"Beschreibung\"],\"O3oNi5\":[\"E-Mail\"],\"ONWvwQ\":[\"Hochladen\"],\"OYHzN1\":[\"Tags\"],\"OaKdb1\":[\"Stammeinheit hinzufügen\"],\"Ocw05N\":[\"Keine gemeinsamen Inhalte\"],\"Oe1ZZ5\":[\"Merkle Batch ID\"],\"P/2ChW\":[\"Vorübergehend\"],\"P2YRxu\":[\"Titel:\"],\"P8fBlG\":[\"Authentifizierung\"],\"POLeLu\":[\"Organisatorischen Umfang zuordnen\"],\"PQKpEq\":[\"Ereignis Hash\"],\"PQs7eb\":[\"URL:\"],\"Pgdahx\":[\"Sie sind offline. Klicken Sie für Details.\"],\"Pi+d4p\":[\"Vorschau konnte nicht geladen werden\"],\"PnzsrT\":[\"Quelltext\"],\"Pt9F7+\":[\"Nicht mit anderen geteilt\"],\"Q+ZWo+\":[\"Höchste Rangnummer, die der Benutzer zuweisen kann (auf 255 für alle, 0 oder leer, um jede Zuweisung zu verhindern). Bereich: 0-255\"],\"Q15keH\":[\"Jetzt aktualisieren\"],\"Q3e3e8\":[\"Dateien erfolgreich hochgeladen!\"],\"QB32Rr\":[\"← Zurück zum Mitarbeiter\"],\"QHZRmq\":[\"Standorte\"],\"QWdKwH\":[\"Verschieben\"],\"Qcag1s\":[\"Widerruf des Zugangs für \",[\"name\"]],\"QgF04t\":[\"Neuer Standort\"],\"QsMBfz\":[\"Ungültige Kombination: min=\",[\"0\"],\", max=0 ergibt keine sichtbaren Arbeitnehmer!\"],\"Qzvlir\":[\"Suche in Beschreibungen...\"],\"R0nC3x\":[\"Abteilung\"],\"RMtipB\":[\"Filter nach Gültigkeit\"],\"RWLG6K\":[\"Zurück zur Startseite\"],\"RXbEvi\":[\"Aktualisiere ...\"],\"Rl5MAb\":[\"Keine übergeordnete Einheit (Stammeinheit)\"],\"RlLl3G\":[\"Aktionen für \",[\"0\"]],\"Rnhhw2\":[\"Mitarbeiter laden...\"],\"S0tQaS\":[\"Zeigen Sie Ihre Kontoinformationen an.\"],\"S3D730\":[\"Organisationseinheit konnte nicht gelöscht werden\"],\"S9pBXd\":[\"Powered by SecPal - a guard's best friend\"],\"SAMXav\":[\"Von Datum\"],\"SBlWEw\":[\"Gemeinsamer Inhalt\"],\"SJmfuf\":[\"Name des Standorts\"],\"SWL25n\":[\"\\\"\",[\"0\"],\"\\\" erfolgreich erstellt\"],\"SlfejT\":[\"Fehler\"],\"Ss8tse\":[\"Der Benutzer kann Ränge sehen (bis \",[\"0\"],\"), die er nicht zuordnen/entfernen kann (bis \",[\"1\"],\").\"],\"Sxm8rQ\":[\"Benutzer\"],\"TUvhZO\":[\"Alles synchronisiert\"],\"TrjzIS\":[\"Zurück zu den Kunden\"],\"Tz0i8g\":[\"Einstellungen\"],\"U3pytU\":[\"Admin\"],\"U8PMKw\":[\"Zeige zwischengespeicherte Daten\"],\"UJmAAK\":[\"Thema\"],\"UMLaSa\":[\"Sie sind offline - Anzeige von Daten im Cache\"],\"UOibiB\":[\"Zusätzliche Eigenschaften\"],\"UPubx6\":[\"Organisatorische Einheit bearbeiten\"],\"URmyfc\":[\"Details\"],\"UXBCwc\":[\"Nachname\"],\"UZEDCT\":[\"Uploads konnten nicht gelöscht werden\"],\"Ua2ccu\":[\"Offline-Benachrichtigung minimieren\"],\"UbRKMZ\":[\"Pending\"],\"UsJkBm\":[\"Anwendung jetzt aktualisieren\"],\"Uv4dnT\":[\"Straße\"],\"UxeTSI\":[\"Hash Chain (Link)\"],\"UyGQAv\":[\"Kunde nicht gefunden\"],\"V4tT03\":[\"Wen kann dieser Benutzer SEHEN?\"],\"VFERo+\":[\"Verwaiste Genesis\"],\"VGMJqV\":[\"Als Geheimnis speichern\"],\"VN4yxm\":[\"Erinnerungen an anstehende Schichten und Aufgaben\"],\"VT+urE\":[\"Suche nach Name oder E-Mail...\"],\"VTfZPy\":[\"Zugang verweigert\"],\"VUEHTd\":[[\"0\",\"plural\",{\"one\":[\"#\",\" Datei bereit zum Hochladen\"],\"other\":[\"#\",\" Dateien bereit zum Hochladen\"]}]],\"Vr7W1t\":[\"Onboarding laden...\"],\"WJ0bmt\":[\"- Lesen: Geheime Details anzeigen\"],\"WNGEr/\":[\"Log Name\"],\"WOF64F\":[\"Benachrichtigungen werden von Ihrem Browser nicht unterstützt. Bitte verwenden Sie einen modernen Browser wie Chrome, Firefox oder Safari.\"],\"WPckwc\":[\"Ungültige oder unsichere URL\"],\"WVwnS9\":[\"Höchste Rangnummer, die der Benutzer sehen kann (255 für alle, 0 oder leer für Nicht-Führungskräfte). Bereich: 0-255\"],\"WXLDBn\":[\"Benutzer oder Rolle auswählen...\"],\"WiWY6F\":[\"Dieses Geheimnis wurde sowohl lokal als auch auf dem Server geändert. Bitte wählen Sie, welche Version Sie behalten möchten.\"],\"Wkumhz\":[\"Keine Einheit ausgewählt\"],\"X1rn/S\":[\"Merkle-Wurzel\"],\"X5DTwE\":[\"Mitarbeiter hinzufügen\"],\"XAS5Dk\":[\"Standort bearbeiten\"],\"XMRrWw\":[\"Status der Verifizierung\"],\"XkkRDx\":[\"Mitarbeiter-Management\"],\"XvjC4F\":[\"Speichern...\"],\"Y/To4Y\":[\"Freiberuflich\"],\"YK/gYI\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Die Organisationseinheit wird dauerhaft entfernt.\"],\"YKrHX4\":[\"HR Zugang\"],\"YXpDzO\":[\"Versuchen Sie, Ihre Filter anzupassen\"],\"YaKsMM\":[\"Erlaubnis:\"],\"YdnZ9S\":[[\"pendingOps\"],\" ausstehende(r) Vorgang(e)\"],\"Yk4+xm\":[\"Anzeigen von \",[\"0\"],\" bis \",[\"1\"],\" von \",[\"2\"],\" Protokollen\"],\"Yop8YX\":[\"Legen Sie eine neue Stammorganisationseinheit an.\"],\"YySIV1\":[\"Benachrichtigungen wurden blockiert. Bitte aktivieren Sie sie in den Einstellungen Ihres Browsers, um wichtige Updates zu erhalten.\"],\"Z3FXyt\":[\"Laden...\"],\"Z8VpDf\":[\"Aktuelle übergeordnete Einheit:\"],\"Z8lGw6\":[\"Teilen\"],\"ZZYTn9\":[\"Formularfelder werden basierend auf dem Schema gerendert\"],\"ZkzhZG\":[\"\\\"\",[\"0\"],\"\\\" erfolgreich aktualisiert\"],\"aG1jjB\":[\"Zur Stammeinheit machen (keine übergeordnete Einheit)\"],\"aK6/iX\":[\"Sie haben keine Berechtigung, auf diese Funktion zuzugreifen. Wenden Sie sich an Ihren Administrator, wenn Sie glauben, dass dies ein Fehler ist.\"],\"aSwfbR\":[\"Einheit\"],\"aVgSDa\":[\"Kundennummer\"],\"aZdNUY\":[\"Team-Nachrichten\"],\"b21I4p\":[\"Neuen Mitarbeiter anlegen\"],\"bG3W+9\":[\"Gültig von\"],\"bS9mQ+\":[\"Offline - wird synchronisiert, wenn online\"],\"c0Kyne\":[\"Mein Profil\"],\"c63pfi\":[\"Sie erhalten jetzt wichtige Updates von SecPal\"],\"c7UClS\":[\"FE\"],\"cJhylo\":[\"Neuer Kunde\"],\"cQbhes\":[\"Neues Geheimnis\"],\"cQlPDN\":[\"Lesen\"],\"cSsOdN\":[\"Minimal sichtbarer rang\"],\"cl2bO8\":[\"Angehängte Dateien\"],\"cnGeoo\":[\"Löschen\"],\"d+F6q9\":[\"Erstellt\"],\"d0J3NP\":[\"Erlaubnisstufen:\"],\"dEgA5A\":[\"Abbrechen\"],\"dJW40d\":[\"Postleitzahl\"],\"dL7xtl\":[\"Benachrichtigungen aktivieren\"],\"dYdPf6\":[[\"label\"],\": N/A\"],\"dbxu+x\":[\"Schritt 1: Konfigurieren der Anzeigeberechtigungen\"],\"dlUnhC\":[\"Zurück zu Mitarbeitern\"],\"doF3ze\":[\"Organisatorischen Umfang bearbeiten\"],\"e/cvV1\":[\"Überprüfen...\"],\"eGHvOp\":[\"Diese Einheit hat \",[\"childCount\"],\" untergeordnete Einheit(en). Bitte löschen oder verschieben Sie die untergeordneten Einheiten, bevor Sie diese Einheit löschen.\"],\"ePK91l\":[\"Bearbeiten\"],\"eeDsml\":[\"Sind Sie sicher, dass Sie \\\"\",[\"0\"],\"\\\" löschen wollen? Diese Aktion kann nicht rückgängig gemacht werden.\"],\"egw0h+\":[\"Schicht-Erinnerungen\"],\"fNnPFj\":[\"Laden von Aktivitätsprotokollen...\"],\"fWCpro\":[\"Letzte Synchronisation: \",[\"0\"]],\"fZHSbC\":[\"Validierungsfehler\"],\"fcWrnU\":[\"Abmelden\"],\"fhKosF\":[\"<0>\",[\"0\"],\"-Feld(er)</0> haben widersprüchliche Änderungen:\"],\"fj5byd\":[\"K.A\"],\"fpzyLj\":[\"Seite \",[\"0\"],\" von \",[\"1\"]],\"ft6Imc\":[\"Geheimnisse durchsuchen\"],\"g82UcF\":[\"Dieser Kunde hat \",[\"0\"],\" Standort(e).\"],\"ginpu1\":[\"Zugriff widerrufen fehlgeschlagen\"],\"goHH65\":[\"Wählen Sie eine Organisationseinheit, um Details anzuzeigen\"],\"gqv5ZL\":[\"Mitarbeiter\"],\"h9bq5K\":[\"Nach Tag filtern\"],\"hXzOVo\":[\"Weiter\"],\"hYgDIe\":[\"Erstellen\"],\"hnFh/R\":[\"Welchen Rangbereich kann dieser Benutzer ZUWEISEN oder ENTFERNEN?\"],\"i0V82m\":[\"Organisationseinheit auswählen\"],\"i0qMbr\":[\"Startseite\"],\"i22E9x\":[\"zurück zu den Standorten\"],\"i4/LY/\":[\"Schreiben\"],\"iB0U5p\":[\"Niedrigste Rangnummer, die der Benutzer vergeben kann (1 = CEO, höhere Zahlen = niedrigere Positionen). Bereich: 1-255\"],\"iH8pgl\":[\"Zurück\"],\"iHlaJP\":[\"Aktivitätsprotokolle aktualisieren\"],\"iJcU0G\":[\"Alle Einheiten\"],\"iVpt+a\":[\"Download des Anhangs fehlgeschlagen\"],\"iW6mc0\":[\"Verwalten Sie Ihre Anwendungspräferenzen.\"],\"iXepfN\":[\"Sind Sie sicher, dass Sie diesen Standort löschen möchten? Diese Aktion kann nicht rückgängig gemacht werden.\"],\"iYX9EM\":[\"Schritt 2: Konfigurieren der Zuweisungsberechtigungen\"],\"itzYLU\":[\"Löschen von \\\"\",[\"unitName\"],\"\\\" nicht möglich\"],\"jEWHqy\":[\"Noch keine Anhänge\"],\"jX9y0M\":[\"Beginnen Sie mit der Erstellung Ihrer ersten Organisationseinheit.\"],\"jesOWC\":[\"Läuft ab (optional):\"],\"jfC/xh\":[\"Kontakt\"],\"joyU4/\":[\"Kein Ablaufdatum\"],\"jpctdh\":[\"Anzeigen\"],\"jqjRKF\":[\"Gültigkeitsdauer\"],\"kL7EQi\":[\"Standort Nummer\"],\"kML1ss\":[\"Aktivieren Sie Benachrichtigungen, um wichtige Updates zu Sicherheitswarnungen, Schichterinnerungen und Systembenachrichtigungen zu erhalten.\"],\"kODvZJ\":[\"Vorname\"],\"kQxFZp\":[\"Wählen Sie Geheim, um Dateien anzuhängen:\"],\"kXWVmG\":[\"Fortschritt der Verarbeitung\"],\"kaKDc0\":[\"Sync-Konflikt erkannt\"],\"kfubA2\":[\"Beenden\"],\"kpt2yf\":[\"SecPal - a guard's best friend\"],\"l75CjT\":[\"Ja\"],\"l7Ja8z\":[\"Keine Seiten gefunden\"],\"l9VeOY\":[\"Teile...\"],\"l9i63I\":[[\"label\"],\": Valid\"],\"lCF0wC\":[\"Aktualisieren\"],\"lGD/gV\":[\"Geheimnisse laden...\"],\"lWndWz\":[\"Organisatorische Einheit anlegen\"],\"lXNHxn\":[\"Text:\"],\"lY19SX\":[\"Keine Führungsposition\"],\"lf0s/G\":[\"Geben Sie dieses Geheimnis für Benutzer oder Rollen frei. Wählen Sie einen Empfänger und eine Berechtigungsstufe aus und legen Sie optional ein Verfallsdatum fest.\"],\"login.checkingSystem\":[\"Überprüfung des Systems...\"],\"login.email\":[\"E-Mail\"],\"login.healthWarning.message\":[\"Das System ist nicht vollständig konfiguriert. Bitte wenden Sie sich an Ihren Administrator.\"],\"login.healthWarning.title\":[\"System nicht bereit\"],\"login.lockedButton\":[\"(\",[\"remainingLockoutSeconds\"],\"s) gesperrt\"],\"login.offlineWarning.message\":[\"Für die Anmeldung ist eine Internetverbindung erforderlich. Bitte überprüfen Sie Ihre Verbindung und versuchen Sie es erneut.\"],\"login.offlineWarning.title\":[\"Keine Internetverbindung\"],\"login.password\":[\"Passwort\"],\"login.rateLimitLocked.message\":[\"Bitte warten Sie \",[\"remainingLockoutSeconds\"],\" Sekunden, bevor Sie es erneut versuchen.\"],\"login.rateLimitLocked.title\":[\"Zu viele gescheiterte Versuche\"],\"login.remainingAttempts\":[[\"remainingAttempts\"],\" verbleibende(r) Versuch(e) vor der vorübergehenden Sperre.\"],\"login.submit\":[\"Einloggen\"],\"login.submitting\":[\"Einloggen...\"],\"login.subtitle\":[\"Your digital guard companion\"],\"login.title\":[\"Einloggen\"],\"lxentK\":[\"Ein unerwarteter Fehler ist aufgetreten\"],\"mAuVsA\":[\"Wiederholungsversuch fehlgeschlagen\"],\"mOFgdC\":[\"Maximum\"],\"mZZN82\":[\"- Admin: Anzeigen + Bearbeiten + Freigeben + Löschen\"],\"mwQODh\":[\"Teilen mit:\"],\"mzI/c+\":[\"Herunterladen\"],\"nQJDF3\":[\"Dateien auswählen\"],\"nVVTqj\":[\"Zuletzt synchronisiert: \",[\"0\"]],\"nWioc3\":[\"<0>Hash Chain:</0> Verifies the sequential integrity of activity logs.<1/><2>Merkle Tree:</2> Batch verification for efficient proof of log inclusion (Security Levels 2-3).<3/><4>OpenTimestamp:</4> Bitcoin blockchain anchoring for immutable proof of existence (Security Level 3).\"],\"nYfwRw\":[\"Datum/Uhrzeit\"],\"ncwQad\":[\"(leer)\"],\"nfagfM\":[\"Valid\"],\"o7jucV\":[\"Der eigene Rang des Benutzers (FE\",[\"userRank\"],\") liegt im sichtbaren Bereich. Sollte dieser Benutzer seine eigenen HR-Daten einsehen/bearbeiten können?\"],\"oJxnEs\":[\"Kunde anlegen\"],\"oLyQSY\":[\"Führungskraft\"],\"oWLjTI\":[\"AGPL v3+\"],\"oXt/Sg\":[\"Hash Chain (Data)\"],\"opYKNZ\":[\"Organisationsstruktur\"],\"ovBPCi\":[\"Standard\"],\"p/78dY\":[\"Position\"],\"p5ku9P\":[\"Der Mindestrang kann nicht größer als der Höchstrang sein.\"],\"pA2odF\":[\"Noch keine Geheimnisse\"],\"pAcLyI\":[\"Organisatorische Einheit\"],\"pK44fh\":[\"Beendet\"],\"peNE68\":[\"Dauerhaft\"],\"pudJRU\":[\"Die Beschreibung darf höchstens 1000 Zeichen lang sein\"],\"qAy4fI\":[\"Keine Qualifikationen zugewiesen\"],\"qIXs9x\":[\"Verursacher\"],\"qJb6G2\":[\"Erneut versuchen\"],\"qVJ3iQ\":[\"Geheimnis erstellen\"],\"qVqcKL\":[\"Selbstzugriffskontrolle erforderlich\"],\"qXoDgO\":[\"Geheimnis bearbeiten\"],\"qe1ces\":[\"Für diesen Dateityp ist keine Vorschau verfügbar.\"],\"quGoru\":[\"Füge ein neues Geheimnis zu deinem Tresor hinzu.\"],\"qwHj/3\":[\"Alle Status\"],\"r6y+jM\":[\"Warnung\"],\"rTaaN1\":[\"Zur Überprüfung einreichen\"],\"rdUucN\":[\"Vorschau\"],\"s+Y9UE\":[\"Zugriffsanweisungen\"],\"sC1Vbh\":[\"Standorte suchen...\"],\"sG1/xZ\":[\"Führungsebene\"],\"ssuwjF\":[\"Niederlassung\"],\"sujToP\":[\"Übergeordnete Einheit\"],\"tJWB4E\":[\"Gewährt von \",[\"0\"],\" am \",[\"1\"]],\"tOv/Bb\":[\"Seite \",[\"currentPage\"],\" von \",[\"totalPages\"],\" (\",[\"0\"],\" \",[\"1\"],\")\"],\"taR1DN\":[\"Zeige zwischengespeicherte Daten. Einige Einheiten können veraltet sein.\"],\"tfDRzk\":[\"Speichern\"],\"uAQUqI\":[\"Status\"],\"uHIPUT\":[\"Synchronisieren...\"],\"uHrWzX\":[\"Verschieben \\\"\",[\"0\"],\"\\\"\"],\"uIxT6Y\":[\"Organisationseinheit konnte nicht verschoben werden\"],\"uJ+Ve2\":[\"Region\"],\"uRTwI2\":[\"Über SecPal\"],\"ucgZ0o\":[\"Organisation\"],\"ufs3xz\":[\"Optionale Beschreibung...\"],\"uhbtDH\":[\"Standort:\"],\"uyJsf6\":[\"Über uns\"],\"v1rRtW\":[\"Test senden\"],\"v9iFOK\":[\"Anhänge laden...\"],\"vERlcd\":[\"Profil\"],\"vFn9Nk\":[\"Präferenzen für Benachrichtigungen\"],\"vHO40l\":[\"Legen Sie eine neue Organisationseinheit unter \\\"\",[\"parentName\"],\"\\\" an.\"],\"vLMns2\":[\"Fehler:\"],\"vLyv1R\":[\"Ausblenden\"],\"vXIe7J\":[\"Sprache\"],\"vXhD8M\":[\"Mitarbeiter laden...\"],\"vmmz7E\":[\"Server-Version\"],\"w66diH\":[\"Teile \\\"\",[\"secretTitle\"],\"\\\"\"],\"wR67is\":[\"Ausstehende Synchronisierung\"],\"wy8uAs\":[\"Entwurf speichern\"],\"xBBUom\":[\"Sind Sie sicher, dass Sie den Zugang für \",[\"name\"],\" widerrufen wollen?\"],\"xCJdfg\":[\"Löschen\"],\"xGRPU0\":[\"<0>Hash-Kette:</0> Überprüft die sequentielle Integrität von Aktivitätsprotokollen. \\\"Ausstehend\\\" zeigt an, dass die Hash-Kette noch aufgebaut wird.<1/><2>Merkle Tree:</2> Batch-Verifizierung für einen effizienten Nachweis der Log-Einbindung (Sicherheitsstufen 2-3). Läuft in der Entwicklung jede Minute, in der Produktion stündlich.<3/><4>OpenTimestamp:</4> Bitcoin-Blockchain-Verankerung für unveränderlichen Existenznachweis (nur Sicherheitsstufe 3). Bitcoin-Bestätigung dauert ~10 Minuten.\"],\"y9c5t9\":[\"Optionale Beschreibung für diese Einheit.\"],\"yBgddv\":[\"Dokumente laden...\"],\"yHzSwP\":[\"← Zurück zu Mitarbeitern\"],\"yLKe22\":[\"Bis heute\"],\"yTjaVJ\":[\"Details zum Aktivitätsprotokoll\"],\"ysdDIA\":[\"Sie sind offline\"],\"yz7wBu\":[\"Schließen\"],\"z78wQd\":[\"Dateien hierher ziehen oder zum Durchsuchen anklicken\"],\"zH+1XO\":[\"Das Löschen von Organisationseinheiten ist im Offline-Modus nicht möglich. Bitte stellen Sie die Verbindung wieder her, um Änderungen vorzunehmen.\"],\"zVyXw4\":[\"Datei zu groß: \",[\"0\"],\". Die maximale Größe beträgt 10MB.\"],\"zga9sT\":[\"OK\"],\"zjo5MA\":[\"Organisationseinheit auswählen...\"],\"zmwvG2\":[\"Telefon\"],\"zzTOEt\":[\"Widerrufen...\"]}");
+/*eslint-disable*/export const messages=JSON.parse("{\"+MTt8f\":[\"\\\"\",[\"unitName\"],\"\\\" löschen?\"],\"+SxetR\":[\"Merkle-Baum\"],\"+b7T3G\":[\"Aktualisiert\"],\"+hZPci\":[\"Datum des Vertragsbeginns\"],\"+jEDxY\":[\"Kunde bearbeiten\"],\"+zy2Nq\":[\"Typ\"],\"/4TFrF\":[\"Vollständiger Name\"],\"/AsMcZ\":[\"Aktiviere ...\"],\"/Q5FOE\":[\"Mitarbeiter bearbeiten\"],\"/YtZc9\":[\"Informationen zur Tätigkeit\"],\"/hW+7s\":[\"Willkommen zum SecPal Onboarding\"],\"036J+m\":[\"Zertifikat:\"],\"0I3xLT\":[\"✓ Benachrichtigungen sind aktiviert. Sie erhalten Updates basierend auf Ihren Einstellungen.\"],\"0VfT1o\":[\"Geheimnis laden...\"],\"0qqeZs\":[\"Um einem Mitarbeiter den Dienstgrad zu entziehen, muss der Benutzer die Berechtigung haben, diesen Dienstgrad zu vergeben. Dies verhindert eine Eskalation der Privilegien.\"],\"0tttqe\":[\"Gültig bis\"],\"0wJVK+\":[\"Grundlegend\"],\"18L2W6\":[\"Mitarbeiter anlegen\"],\"18zG7E\":[\"Verwalten Sie Ihre internen Organisationseinheiten wie Abteilungen, Niederlassungen und Teams.\"],\"1A3EXy\":[\"Ausklappen\"],\"1DBGsz\":[\"Anmerkungen\"],\"1GixfT\":[\"Bald ablaufend\"],\"1eiedP\":[\"Organisatorische Struktur\"],\"1ep1s+\":[\"Zuletzt geändert: \",[\"0\"]],\"1k5CRu\":[\"Unterstützt: Bilder, PDFs, Dokumente (max. 10MB)\"],\"1oL0IJ\":[\"Sind Sie sicher, dass Sie diesen Anhang löschen möchten?\"],\"2BBAbc\":[\"Liste\"],\"2GT3Hf\":[\"Global\"],\"2wgink\":[\"Anzeigen von <0>\",[\"0\"],\"</0> bis <1>\",[\"1\"],\"</1> von <2>\",[\"2\"],\"</2> Mitarbeitern\"],\"3T71mA\":[\"Wählen Sie, welche Benachrichtigungen Sie erhalten möchten\"],\"3TgTKY\":[\"Datei konnte nicht gelöscht werden\"],\"3aPOvd\":[\"Anhang konnte nicht gelöscht werden\"],\"3dwitg\":[\"Mitteilungen von Teammitgliedern und Vorgesetzten\"],\"3e8eQb\":[\"Keine organisatorischen Einheiten\"],\"3wV73y\":[\"Stadt\"],\"43Ygvu\":[\"Fehler beim Laden des Geheimnisses\"],\"48oQna\":[\"Das Anlegen von Organisationseinheiten ist im Offline-Modus nicht möglich. Bitte stellen Sie die Verbindung wieder her, um Änderungen vorzunehmen.\"],\"4AY++3\":[\"Standort-Nummer:\"],\"4DZJcB\":[\"Die Bearbeitung von Organisationseinheiten ist im Offline-Modus nicht möglich. Bitte stellen Sie die Verbindung wieder her, um Änderungen vorzunehmen.\"],\"4DjwT6\":[\"Legen Sie fest, welchen Rangbereich dieser Benutzer sehen kann. Lassen Sie das Feld leer oder setzen Sie es auf 0, um nur Mitarbeiter ohne Führungsaufgaben (Sicherheitsmitarbeiter) zu sehen.\"],\"4FcOp+\":[\"Fehler beim Laden eines Mitarbeiters\"],\"4OaTD4\":[\"in Freistellung\"],\"4Z4JSO\":[\"Name des Kunden\"],\"4cxQxp\":[\"Anzeigen von <0>\",[\"0\"],\"</0> bis <1>\",[\"1\"],\"</1> von <2>\",[\"2\"],\"</2> Mitarbeitern\"],\"4fjnRg\":[\"-- Wählen Sie ein Geheimnis --\"],\"4geA9Q\":[\"Fehler beim Laden von Geheimnissen\"],\"4yI3Vk\":[\"Zurück zu Geheimnisse\"],\"4yv9An\":[\"Erweitert\"],\"5/FzWg\":[\"Gelöscht \",[\"count\"],\" abgeschlossene(r) Upload(s)\"],\"58X1WA\":[\"Alle Einheiten\"],\"5A0Vat\":[\"Untergeordnete Einheit hinzufügen\"],\"5HYWB0\":[\"Bereich\"],\"5I+k35\":[\"Mitarbeiter Nummer\"],\"5NZpX8\":[\"Listenansicht\"],\"5ZulCG\":[\"Typ ist erforderlich\"],\"5cVSqz\":[\"Gemeinsam genutzt mit (\",[\"0\"],\")\"],\"5dJK4M\":[\"Rollen\"],\"5eHyY+\":[\"Test-Benachrichtigung\"],\"5fap5A\":[\"Standort nicht gefunden\"],\"5gqNQl\":[\"Gitternetz-Ansicht\"],\"5lq7gm\":[\"Standardmäßig (nicht aktiviert) können Benutzer ihre eigen Personalakte weder sehen noch bearbeiten. Dies verhindert die Selbstmanipulation von Gehalt, Rang usw. Aktivieren Sie dieses Kontrollkästchen nur, wenn Sie möchten, dass dieser Benutzer auf seine eigene Personalakte zugreifen kann.\"],\"5tmzAO\":[\"Die neue Einheit wird unter dieser übergeordneten Einheit erstellt.\"],\"6GBt0m\":[\"Metadaten\"],\"6LhVHS\":[\"Übergeordnete Einheit\"],\"6Q2qjr\":[\"Kunden suchen...\"],\"6VfR2o\":[\"Qualifikationen\"],\"6YtxFj\":[\"Name\"],\"6gRgw8\":[\"Wiederholen\"],\"6kPk3+\":[\"Persönliche Informationen\"],\"6splsl\":[\"Ungültige Kombination: Es werden keine Mitarbeiter mit diesem Bereich angezeigt.\"],\"6yvfgr\":[\"Benachrichtigungen aktiviert\"],\"7/EICV\":[\"- Schreiben: Ansehen + Bearbeiten Geheimnis\"],\"7/tqYA\":[\"neue übergeordnete Einheit auswählen\"],\"7Att6/\":[\"Qualifikationen laden...\"],\"7L01XJ\":[\"Aktionen\"],\"7OWvhx\":[\"OpenTimestamp Bestätigt\"],\"7TJx4N\":[\"Eine neue Version von SecPal ist verfügbar.\"],\"7i8j3G\":[\"Unternehmen\"],\"7sNhEz\":[\"Benutzername\"],\"7uHWR+\":[\"Geburtsdatum\"],\"8/brI5\":[\"Name ist erforderlich\"],\"842lnw\":[\"Erstellen Sie Ihr erstes Geheimnis, um loszulegen\"],\"84T3OZ\":[\"Maximal zuordenbarer Rang\"],\"876pfE\":[\"Kunde\"],\"89yiwg\":[\"Der Benutzer kann KEINEN Rang zuweisen ODER entfernen (verhindert eine Eskalation der Rechte).\"],\"8CnLST\":[\"Geheimnisse ansehen\"],\"8RYNR1\":[\"Daten aktualisieren\"],\"8Tg/JR\":[\"Benutzerdefiniert\"],\"8ZsakT\":[\"Passwort\"],\"8bl30j\":[\"Aktualisieren Sie die Details der Organisationseinheit.\"],\"8rE61T\":[\"Rechnungsadresse\"],\"8vETh9\":[\"Anzeigen\"],\"8z4jWY\":[\"Sicherheitswarnungen\"],\"9AcM8x\":[\"Löschen Abgeschlossen\"],\"9H0JwN\":[\"Mitarbeiter #\"],\"9P/b/5\":[\"App-Updates und Wartungsbenachrichtigungen\"],\"9bEGEb\":[\"Läuft ab:\"],\"9cDpsw\":[\"Berechtigungen\"],\"9cRAwD\":[\"← Zurück zu Geheimnissen\"],\"A1taO8\":[\"Suche\"],\"A8qnCR\":[\"Dies ist eine Testbenachrichtigung von SecPal\"],\"AAnV/D\":[\"Wählen Sie eine neue übergeordnete Stelle für diese Organisationseinheit aus.\"],\"AN9mgS\":[[\"filename\"],\" aus der Warteschlange entfernt\"],\"AbIhll\":[\"Vertragsart\"],\"AuwJQ1\":[\"Ungültiger Dateityp: \",[\"0\"],\". Nur Bilder, PDFs und Dokumente sind erlaubt.\"],\"BHoh+C\":[\"Minimaler zuweisbarer Rang\"],\"BLXWJv\":[\"Keine Kunden gefunden\"],\"Baxs4G\":[\"Standort löschen\"],\"BfQvQ5\":[\"System-Updates\"],\"Bgjf1M\":[\"Geheimnis teilen fehlgeschlagen\"],\"BgqIX/\":[\"OpenTimestamp: Nicht zutreffend\"],\"BltQR1\":[\"Standorte anzeigen\"],\"C+ffQE\":[\"Vollzeit\"],\"C448X5\":[\"z.B. Niederlassung Berlin\"],\"CCXa60\":[\"Keine Dokumente hochgeladen\"],\"CEwoiy\":[\"Niedrigste Rangnummer, die der Benutzer sehen kann (1 = CEO, höhere Zahlen = niedrigere Positionen). Bereich: 1-255\"],\"CXDHcv\":[\"Raster\"],\"CjXztc\":[\"Zugangskontrolle\"],\"ClD8Z1\":[\"Details zur Beschäftigung\"],\"Co/eKl\":[\"Benutzer können ihre eigenen HR-Daten anzeigen/bearbeiten\"],\"Cyc2u5\":[\"Aktivitätsprotokollinformationen und Überprüfungsstatus anzeigen\"],\"CzsL7L\":[\"Teilzeit\"],\"D+NlUC\":[\"System\"],\"D2aVgL\":[\"Vorheriger Hash\"],\"D5Y6ym\":[\"Schritt 3: Kontrolle des Selbstzugriffs\"],\"DBoLwV\":[\"Alle Tags\"],\"DBr12o\":[\"Anhänge (\",[\"0\"],\")\"],\"DDffEz\":[\"Maximal sichtbarer Rang\"],\"DHhJ7s\":[\"Vorherige\"],\"DYlwPh\":[\"Kontaktperson\"],\"DebAtR\":[[\"errorOps\"],\" Operation(en) fehlgeschlagen\"],\"Du6bPw\":[\"Adresse\"],\"E/muDO\":[\"Dokumente\"],\"EF2EU9\":[\"Löschen...\"],\"EH48dY\":[\"Einige Funktionen sind möglicherweise eingeschränkt. Ihre Änderungen werden synchronisiert, sobald Sie wieder online sind.\"],\"ELI7SU\":[[\"label\"],\": Ausstehend\"],\"ELa85n\":[\"Alle Typen\"],\"EOv8gS\":[\"Koordinaten\"],\"EPp7ex\":[\"Der Name darf maximal 255 Zeichen lang sein\"],\"EkH9pt\":[\"Update\"],\"ErGYe4\":[\"Aktualisieren Sie die Angaben zu Ihrem Geheimnis.\"],\"F6pfE9\":[\"Aktiv\"],\"F7ErMS\":[[\"label\"],\": Ungültig\"],\"FEBEYO\":[\"Geheimnisse durchsuchen...\"],\"FQBaXG\":[\"Aktivieren\"],\"FQfoqR\":[\"Ihre lokale Version\"],\"FSimqX\":[\"Kunde auswählen...\"],\"FYzHRF\":[\"Holding\"],\"Failed to change language. Please try again.\":[\"Die Sprache konnte nicht geändert werden. Bitte versuchen Sie es erneut.\"],\"FjhOTo\":[\"Kunde löschen\"],\"Flmwkt\":[\"Mitarbeiter\"],\"FtJ+Sn\":[\"Meine Organisation\"],\"FxAKnZ\":[\"Kritische Sicherheitsbenachrichtigungen und -warnungen\"],\"GW+FsP\":[\"Auswahl anwenden\"],\"GXsAby\":[\"Widerrufen\"],\"GhMjr9\":[\"Dateien verschlüsseln...\"],\"Gwmfl9\":[\"Detailfenster schließen\"],\"GxkJXS\":[\"Hochladen...\"],\"GyKiE+\":[\"Untergeordnete Einheiten müssen gelöscht oder einer anderen übergeordneten Einheit neu zugeordnet werden, bevor diese Organisationseinheit entfernt werden kann.\"],\"GyrqoW\":[\"Aktivitätsprotokolle\"],\"H86f9p\":[\"Zusammenklappen\"],\"H8BDKE\":[\"Zurück zur Liste\"],\"HBk9Vi\":[\"neue übergeordnete Einheit\"],\"HM56Bx\":[\"Erstellen...\"],\"HVUgrA\":[\"Die Bereichszuweisung konnte nicht gespeichert werden.\"],\"IUwGEM\":[\"Änderungen speichern\"],\"IXHBs4\":[\"Standort anlegen\"],\"IagCbF\":[\"URL\"],\"Id+qAX\":[\"Keine Aktivitätsprotokolle gefunden.\"],\"IiFHaM\":[\"Wählen Sie Ihre bevorzugte Sprache für die Anwendung.\"],\"Iv4GiC\":[\"Der Typ bestimmt die Rolle der Einheit in der Hierarchie.\"],\"IvB46R\":[\"Zur Ansicht herunterladen\"],\"J00gME\":[\"Anzeigen von <0>\",[\"0\"],\"</0> bis <1>\",[\"1\"],\"</1> von <2>\",[\"2\"],\"</2> Standorten\"],\"J4cXlI\":[\"OpenTimestamp\"],\"JHB/Ky\":[\"Invalid\"],\"JLJLL1\":[\"Untergeordnete Einheit hinzufügen\"],\"JNCzPW\":[\"Land\"],\"JQEpPE\":[\"Sicherheitsstufe\"],\"JRVmaF\":[\"Hochgeladen: \",[\"0\"]],\"Ju2mNv\":[\"Keine Geheimnisse gefunden\"],\"K0pe4j\":[\"Alle Protokolle\"],\"K7P0jz\":[\"Zuletzt aktualisiert\"],\"K8Qnlj\":[\"Verschiebe...\"],\"KELjcQ\":[\"Willkommen zu SecPal\"],\"Kfa+gn\":[\"Vorvertrag\"],\"KhU0au\":[\"Nur offline\"],\"KnN1Tu\":[\"Läuft ab\"],\"Kzoh+W\":[\"SecPal - der beste Freund eines Wachmanns. Eine Offline-First Progressive Web App für Sicherheitspersonal, die digitale Wachbücher mit modernem Service-Management verbindet.\"],\"LCDse8\":[\"Automatisches Aktualisieren (30s)\"],\"LII+b3\":[\"Wiederholungsversuch geplant. Dateien werden hochgeladen, wenn das Netzwerk verfügbar ist.\"],\"Ln8nFJ\":[\"Ziel teilen\"],\"LoUIHT\":[\"Minijob\"],\"LtI9AS\":[\"Eigentümer\"],\"M1RnFv\":[\"Abgelaufen\"],\"M3sRlm\":[\"Fehler beim Laden von Mitarbeitern\"],\"M5lRtC\":[\"Läuft ab: \",[\"0\"]],\"MAv87s\":[\"Geheimnisse\"],\"MLW/Q/\":[\"Das Verschieben von Organisationseinheiten ist im Offline-Modus nicht möglich. Bitte stellen Sie die Verbindung wieder her, um Änderungen vorzunehmen.\"],\"MpF6Ix\":[\"Upload-Warteschlange\"],\"MvsnLi\":[\"Upload-Wiederholungsversuch konnte nicht geplant werden\"],\"MyWb0G\":[\"Verwalten Sie Ihre Geheimnisse und sensiblen Informationen auf sichere Weise.\"],\"N40H+G\":[\"Alle\"],\"NgNAlL\":[\"Schritt \",[\"currentStep\"],\" von \",[\"totalSteps\"]],\"NihQNk\":[\"Kunden\"],\"NoNwIX\":[\"Inaktiv\"],\"Np2enW\":[\"Wählen Sie eine neue übergeordnete Einheit oder machen Sie diese zu einer Stammeinheit.\"],\"NtEny3\":[\"Verschlüsselungsschlüssel für Anhänge kann nicht geladen werden. Bitte aktualisieren Sie die Seite oder kontaktieren Sie den Support, falls das Problem weiterhin besteht.\"],\"Nu4DdT\":[\"Synchronisieren\"],\"Nu4oKW\":[\"Beschreibung\"],\"O3oNi5\":[\"E-Mail\"],\"ONWvwQ\":[\"Hochladen\"],\"OYHzN1\":[\"Tags\"],\"OaKdb1\":[\"Stammeinheit hinzufügen\"],\"Ocw05N\":[\"Keine gemeinsamen Inhalte\"],\"Oe1ZZ5\":[\"Merkle Batch ID\"],\"P/2ChW\":[\"Vorübergehend\"],\"P2YRxu\":[\"Titel:\"],\"P8fBlG\":[\"Authentifizierung\"],\"POLeLu\":[\"Organisatorischen Umfang zuordnen\"],\"PQKpEq\":[\"Ereignis Hash\"],\"PQs7eb\":[\"URL:\"],\"Pgdahx\":[\"Sie sind offline. Klicken Sie für Details.\"],\"Pi+d4p\":[\"Vorschau konnte nicht geladen werden\"],\"PnzsrT\":[\"Quelltext\"],\"Pt9F7+\":[\"Nicht mit anderen geteilt\"],\"Q+ZWo+\":[\"Höchste Rangnummer, die der Benutzer zuweisen kann (auf 255 für alle, 0 oder leer, um jede Zuweisung zu verhindern). Bereich: 0-255\"],\"Q15keH\":[\"Jetzt aktualisieren\"],\"Q3e3e8\":[\"Dateien erfolgreich hochgeladen!\"],\"QB32Rr\":[\"← Zurück zum Mitarbeiter\"],\"QHZRmq\":[\"Standorte\"],\"QWdKwH\":[\"Verschieben\"],\"Qcag1s\":[\"Widerruf des Zugangs für \",[\"name\"]],\"QgF04t\":[\"Neuer Standort\"],\"QsMBfz\":[\"Ungültige Kombination: min=\",[\"0\"],\", max=0 ergibt keine sichtbaren Arbeitnehmer!\"],\"Qzvlir\":[\"Suche in Beschreibungen...\"],\"R0nC3x\":[\"Abteilung\"],\"RMtipB\":[\"Filter nach Gültigkeit\"],\"RWLG6K\":[\"Zurück zur Startseite\"],\"RXbEvi\":[\"Aktualisiere ...\"],\"Rl5MAb\":[\"Keine übergeordnete Einheit (Stammeinheit)\"],\"RlLl3G\":[\"Aktionen für \",[\"0\"]],\"Rnhhw2\":[\"Mitarbeiter laden...\"],\"S0tQaS\":[\"Zeigen Sie Ihre Kontoinformationen an.\"],\"S3D730\":[\"Organisationseinheit konnte nicht gelöscht werden\"],\"S9pBXd\":[\"Powered by SecPal - a guard's best friend\"],\"SAMXav\":[\"Von Datum\"],\"SBlWEw\":[\"Gemeinsamer Inhalt\"],\"SJmfuf\":[\"Name des Standorts\"],\"SWL25n\":[\"\\\"\",[\"0\"],\"\\\" erfolgreich erstellt\"],\"SlfejT\":[\"Fehler\"],\"Ss8tse\":[\"Der Benutzer kann Ränge sehen (bis \",[\"0\"],\"), die er nicht zuordnen/entfernen kann (bis \",[\"1\"],\").\"],\"Sxm8rQ\":[\"Benutzer\"],\"TUvhZO\":[\"Alles synchronisiert\"],\"TrjzIS\":[\"Zurück zu den Kunden\"],\"Tz0i8g\":[\"Einstellungen\"],\"U3pytU\":[\"Admin\"],\"U8PMKw\":[\"Zeige zwischengespeicherte Daten\"],\"UJmAAK\":[\"Thema\"],\"UMLaSa\":[\"Sie sind offline - Anzeige von Daten im Cache\"],\"UOibiB\":[\"Zusätzliche Eigenschaften\"],\"UPubx6\":[\"Organisatorische Einheit bearbeiten\"],\"URmyfc\":[\"Details\"],\"UXBCwc\":[\"Nachname\"],\"UZEDCT\":[\"Uploads konnten nicht gelöscht werden\"],\"Ua2ccu\":[\"Offline-Benachrichtigung minimieren\"],\"UbRKMZ\":[\"Pending\"],\"UsJkBm\":[\"Anwendung jetzt aktualisieren\"],\"Uv4dnT\":[\"Straße\"],\"UxeTSI\":[\"Hash-Kette (Link)\"],\"UyGQAv\":[\"Kunde nicht gefunden\"],\"V4tT03\":[\"Wen kann dieser Benutzer SEHEN?\"],\"VFERo+\":[\"Verwaiste Genesis\"],\"VGMJqV\":[\"Als Geheimnis speichern\"],\"VN4yxm\":[\"Erinnerungen an anstehende Schichten und Aufgaben\"],\"VT+urE\":[\"Suche nach Name oder E-Mail...\"],\"VTfZPy\":[\"Zugang verweigert\"],\"VUEHTd\":[[\"0\",\"plural\",{\"one\":[\"#\",\" Datei bereit zum Hochladen\"],\"other\":[\"#\",\" Dateien bereit zum Hochladen\"]}]],\"Vr7W1t\":[\"Onboarding laden...\"],\"WJ0bmt\":[\"- Lesen: Geheime Details anzeigen\"],\"WNGEr/\":[\"Log Name\"],\"WOF64F\":[\"Benachrichtigungen werden von Ihrem Browser nicht unterstützt. Bitte verwenden Sie einen modernen Browser wie Chrome, Firefox oder Safari.\"],\"WPckwc\":[\"Ungültige oder unsichere URL\"],\"WVwnS9\":[\"Höchste Rangnummer, die der Benutzer sehen kann (255 für alle, 0 oder leer für Nicht-Führungskräfte). Bereich: 0-255\"],\"WXLDBn\":[\"Benutzer oder Rolle auswählen...\"],\"WiWY6F\":[\"Dieses Geheimnis wurde sowohl lokal als auch auf dem Server geändert. Bitte wählen Sie, welche Version Sie behalten möchten.\"],\"Wkumhz\":[\"Keine Einheit ausgewählt\"],\"X1rn/S\":[\"Merkle-Wurzel\"],\"X5DTwE\":[\"Mitarbeiter hinzufügen\"],\"XAS5Dk\":[\"Standort bearbeiten\"],\"XMRrWw\":[\"Status der Verifizierung\"],\"XkkRDx\":[\"Mitarbeiter-Management\"],\"XvjC4F\":[\"Speichern...\"],\"Y/To4Y\":[\"Freiberuflich\"],\"YK/gYI\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Die Organisationseinheit wird dauerhaft entfernt.\"],\"YKrHX4\":[\"HR Zugang\"],\"YXpDzO\":[\"Versuchen Sie, Ihre Filter anzupassen\"],\"YaKsMM\":[\"Erlaubnis:\"],\"YdnZ9S\":[[\"pendingOps\"],\" ausstehende(r) Vorgang(e)\"],\"Yk4+xm\":[\"Anzeigen von \",[\"0\"],\" bis \",[\"1\"],\" von \",[\"2\"],\" Protokollen\"],\"Yop8YX\":[\"Legen Sie eine neue Stammorganisationseinheit an.\"],\"YySIV1\":[\"Benachrichtigungen wurden blockiert. Bitte aktivieren Sie sie in den Einstellungen Ihres Browsers, um wichtige Updates zu erhalten.\"],\"Z3FXyt\":[\"Laden...\"],\"Z8VpDf\":[\"Aktuelle übergeordnete Einheit:\"],\"Z8lGw6\":[\"Teilen\"],\"ZZYTn9\":[\"Formularfelder werden basierend auf dem Schema gerendert\"],\"ZkzhZG\":[\"\\\"\",[\"0\"],\"\\\" erfolgreich aktualisiert\"],\"aG1jjB\":[\"Zur Stammeinheit machen (keine übergeordnete Einheit)\"],\"aK6/iX\":[\"Sie haben keine Berechtigung, auf diese Funktion zuzugreifen. Wenden Sie sich an Ihren Administrator, wenn Sie glauben, dass dies ein Fehler ist.\"],\"aSwfbR\":[\"Einheit\"],\"aVgSDa\":[\"Kundennummer\"],\"aZdNUY\":[\"Team-Nachrichten\"],\"b21I4p\":[\"Neuen Mitarbeiter anlegen\"],\"bG3W+9\":[\"Gültig von\"],\"bS9mQ+\":[\"Offline - wird synchronisiert, wenn online\"],\"c0Kyne\":[\"Mein Profil\"],\"c63pfi\":[\"Sie erhalten jetzt wichtige Updates von SecPal\"],\"c7UClS\":[\"FE\"],\"cJhylo\":[\"Neuer Kunde\"],\"cQbhes\":[\"Neues Geheimnis\"],\"cQlPDN\":[\"Lesen\"],\"cSsOdN\":[\"Minimal sichtbarer rang\"],\"cl2bO8\":[\"Angehängte Dateien\"],\"cnGeoo\":[\"Löschen\"],\"d+F6q9\":[\"Erstellt\"],\"d0J3NP\":[\"Erlaubnisstufen:\"],\"dEgA5A\":[\"Abbrechen\"],\"dJW40d\":[\"Postleitzahl\"],\"dL7xtl\":[\"Benachrichtigungen aktivieren\"],\"dYdPf6\":[[\"label\"],\": Nicht zutreffend\"],\"dbxu+x\":[\"Schritt 1: Konfigurieren der Anzeigeberechtigungen\"],\"dlUnhC\":[\"Zurück zu Mitarbeitern\"],\"doF3ze\":[\"Organisatorischen Umfang bearbeiten\"],\"e/cvV1\":[\"Überprüfen...\"],\"eGHvOp\":[\"Diese Einheit hat \",[\"childCount\"],\" untergeordnete Einheit(en). Bitte löschen oder verschieben Sie die untergeordneten Einheiten, bevor Sie diese Einheit löschen.\"],\"ePK91l\":[\"Bearbeiten\"],\"eeDsml\":[\"Sind Sie sicher, dass Sie \\\"\",[\"0\"],\"\\\" löschen wollen? Diese Aktion kann nicht rückgängig gemacht werden.\"],\"egw0h+\":[\"Schicht-Erinnerungen\"],\"fNnPFj\":[\"Laden von Aktivitätsprotokollen...\"],\"fWCpro\":[\"Letzte Synchronisation: \",[\"0\"]],\"fZHSbC\":[\"Validierungsfehler\"],\"fcWrnU\":[\"Abmelden\"],\"fhKosF\":[\"<0>\",[\"0\"],\"-Feld(er)</0> haben widersprüchliche Änderungen:\"],\"fj5byd\":[\"K.A\"],\"fpzyLj\":[\"Seite \",[\"0\"],\" von \",[\"1\"]],\"ft6Imc\":[\"Geheimnisse durchsuchen\"],\"g82UcF\":[\"Dieser Kunde hat \",[\"0\"],\" Standort(e).\"],\"ginpu1\":[\"Zugriff widerrufen fehlgeschlagen\"],\"goHH65\":[\"Wählen Sie eine Organisationseinheit, um Details anzuzeigen\"],\"gqv5ZL\":[\"Mitarbeiter\"],\"h9bq5K\":[\"Nach Tag filtern\"],\"hXzOVo\":[\"Weiter\"],\"hYgDIe\":[\"Erstellen\"],\"hnFh/R\":[\"Welchen Rangbereich kann dieser Benutzer ZUWEISEN oder ENTFERNEN?\"],\"i0V82m\":[\"Organisationseinheit auswählen\"],\"i0qMbr\":[\"Startseite\"],\"i22E9x\":[\"zurück zu den Standorten\"],\"i4/LY/\":[\"Schreiben\"],\"iB0U5p\":[\"Niedrigste Rangnummer, die der Benutzer vergeben kann (1 = CEO, höhere Zahlen = niedrigere Positionen). Bereich: 1-255\"],\"iH8pgl\":[\"Zurück\"],\"iHlaJP\":[\"Aktivitätsprotokolle aktualisieren\"],\"iJcU0G\":[\"Alle Einheiten\"],\"iVpt+a\":[\"Download des Anhangs fehlgeschlagen\"],\"iW6mc0\":[\"Verwalten Sie Ihre Anwendungspräferenzen.\"],\"iXepfN\":[\"Sind Sie sicher, dass Sie diesen Standort löschen möchten? Diese Aktion kann nicht rückgängig gemacht werden.\"],\"iYX9EM\":[\"Schritt 2: Konfigurieren der Zuweisungsberechtigungen\"],\"itzYLU\":[\"Löschen von \\\"\",[\"unitName\"],\"\\\" nicht möglich\"],\"jEWHqy\":[\"Noch keine Anhänge\"],\"jX9y0M\":[\"Beginnen Sie mit der Erstellung Ihrer ersten Organisationseinheit.\"],\"jesOWC\":[\"Läuft ab (optional):\"],\"jfC/xh\":[\"Kontakt\"],\"joyU4/\":[\"Kein Ablaufdatum\"],\"jpctdh\":[\"Anzeigen\"],\"jqjRKF\":[\"Gültigkeitsdauer\"],\"kL7EQi\":[\"Standort Nummer\"],\"kML1ss\":[\"Aktivieren Sie Benachrichtigungen, um wichtige Updates zu Sicherheitswarnungen, Schichterinnerungen und Systembenachrichtigungen zu erhalten.\"],\"kODvZJ\":[\"Vorname\"],\"kQxFZp\":[\"Wählen Sie Geheim, um Dateien anzuhängen:\"],\"kXWVmG\":[\"Fortschritt der Verarbeitung\"],\"kaKDc0\":[\"Sync-Konflikt erkannt\"],\"kfubA2\":[\"Beenden\"],\"kpt2yf\":[\"SecPal - a guard's best friend\"],\"l75CjT\":[\"Ja\"],\"l7Ja8z\":[\"Keine Seiten gefunden\"],\"l9VeOY\":[\"Teile...\"],\"l9i63I\":[[\"label\"],\": Gültig\"],\"lCF0wC\":[\"Aktualisieren\"],\"lGD/gV\":[\"Geheimnisse laden...\"],\"lWndWz\":[\"Organisatorische Einheit anlegen\"],\"lXNHxn\":[\"Text:\"],\"lY19SX\":[\"Keine Führungsposition\"],\"lf0s/G\":[\"Geben Sie dieses Geheimnis für Benutzer oder Rollen frei. Wählen Sie einen Empfänger und eine Berechtigungsstufe aus und legen Sie optional ein Verfallsdatum fest.\"],\"login.checkingSystem\":[\"Überprüfung des Systems...\"],\"login.email\":[\"E-Mail\"],\"login.healthWarning.message\":[\"Das System ist nicht vollständig konfiguriert. Bitte wenden Sie sich an Ihren Administrator.\"],\"login.healthWarning.title\":[\"System nicht bereit\"],\"login.lockedButton\":[\"(\",[\"remainingLockoutSeconds\"],\"s) gesperrt\"],\"login.offlineWarning.message\":[\"Für die Anmeldung ist eine Internetverbindung erforderlich. Bitte überprüfen Sie Ihre Verbindung und versuchen Sie es erneut.\"],\"login.offlineWarning.title\":[\"Keine Internetverbindung\"],\"login.password\":[\"Passwort\"],\"login.rateLimitLocked.message\":[\"Bitte warten Sie \",[\"remainingLockoutSeconds\"],\" Sekunden, bevor Sie es erneut versuchen.\"],\"login.rateLimitLocked.title\":[\"Zu viele gescheiterte Versuche\"],\"login.remainingAttempts\":[[\"remainingAttempts\"],\" verbleibende(r) Versuch(e) vor der vorübergehenden Sperre.\"],\"login.submit\":[\"Einloggen\"],\"login.submitting\":[\"Einloggen...\"],\"login.subtitle\":[\"Your digital guard companion\"],\"login.title\":[\"Einloggen\"],\"lxentK\":[\"Ein unerwarteter Fehler ist aufgetreten\"],\"mAuVsA\":[\"Wiederholungsversuch fehlgeschlagen\"],\"mOFgdC\":[\"Maximum\"],\"mZZN82\":[\"- Admin: Anzeigen + Bearbeiten + Freigeben + Löschen\"],\"mwQODh\":[\"Teilen mit:\"],\"mzI/c+\":[\"Herunterladen\"],\"nQJDF3\":[\"Dateien auswählen\"],\"nVVTqj\":[\"Zuletzt synchronisiert: \",[\"0\"]],\"nWioc3\":[\"<0>Hash Chain:</0> Verifies the sequential integrity of activity logs.<1/><2>Merkle Tree:</2> Batch verification for efficient proof of log inclusion (Security Levels 2-3).<3/><4>OpenTimestamp:</4> Bitcoin blockchain anchoring for immutable proof of existence (Security Level 3).\"],\"nYfwRw\":[\"Datum/Uhrzeit\"],\"ncwQad\":[\"(leer)\"],\"nfagfM\":[\"Valid\"],\"o7jucV\":[\"Der eigene Rang des Benutzers (FE\",[\"userRank\"],\") liegt im sichtbaren Bereich. Sollte dieser Benutzer seine eigenen HR-Daten einsehen/bearbeiten können?\"],\"oJxnEs\":[\"Kunde anlegen\"],\"oLyQSY\":[\"Führungskraft\"],\"oWLjTI\":[\"AGPL v3+\"],\"oXt/Sg\":[\"Hash-Kette (Daten)\"],\"opYKNZ\":[\"Organisationsstruktur\"],\"ovBPCi\":[\"Standard\"],\"p/78dY\":[\"Position\"],\"p5ku9P\":[\"Der Mindestrang kann nicht größer als der Höchstrang sein.\"],\"pA2odF\":[\"Noch keine Geheimnisse\"],\"pAcLyI\":[\"Organisatorische Einheit\"],\"pK44fh\":[\"Beendet\"],\"peNE68\":[\"Dauerhaft\"],\"pudJRU\":[\"Die Beschreibung darf höchstens 1000 Zeichen lang sein\"],\"qAy4fI\":[\"Keine Qualifikationen zugewiesen\"],\"qIXs9x\":[\"Verursacher\"],\"qJb6G2\":[\"Erneut versuchen\"],\"qVJ3iQ\":[\"Geheimnis erstellen\"],\"qVqcKL\":[\"Selbstzugriffskontrolle erforderlich\"],\"qXoDgO\":[\"Geheimnis bearbeiten\"],\"qe1ces\":[\"Für diesen Dateityp ist keine Vorschau verfügbar.\"],\"quGoru\":[\"Füge ein neues Geheimnis zu deinem Tresor hinzu.\"],\"qwHj/3\":[\"Alle Status\"],\"r6y+jM\":[\"Warnung\"],\"rTaaN1\":[\"Zur Überprüfung einreichen\"],\"rdUucN\":[\"Vorschau\"],\"s+Y9UE\":[\"Zugriffsanweisungen\"],\"sC1Vbh\":[\"Standorte suchen...\"],\"sG1/xZ\":[\"Führungsebene\"],\"ssuwjF\":[\"Niederlassung\"],\"sujToP\":[\"Übergeordnete Einheit\"],\"tJWB4E\":[\"Gewährt von \",[\"0\"],\" am \",[\"1\"]],\"tOv/Bb\":[\"Seite \",[\"currentPage\"],\" von \",[\"totalPages\"],\" (\",[\"0\"],\" \",[\"1\"],\")\"],\"taR1DN\":[\"Zeige zwischengespeicherte Daten. Einige Einheiten können veraltet sein.\"],\"tfDRzk\":[\"Speichern\"],\"uAQUqI\":[\"Status\"],\"uHIPUT\":[\"Synchronisieren...\"],\"uHrWzX\":[\"Verschieben \\\"\",[\"0\"],\"\\\"\"],\"uIxT6Y\":[\"Organisationseinheit konnte nicht verschoben werden\"],\"uJ+Ve2\":[\"Region\"],\"uRTwI2\":[\"Über SecPal\"],\"ucgZ0o\":[\"Organisation\"],\"ufs3xz\":[\"Optionale Beschreibung...\"],\"uhbtDH\":[\"Standort:\"],\"uyJsf6\":[\"Über uns\"],\"v1rRtW\":[\"Test senden\"],\"v9iFOK\":[\"Anhänge laden...\"],\"vERlcd\":[\"Profil\"],\"vFn9Nk\":[\"Präferenzen für Benachrichtigungen\"],\"vHO40l\":[\"Legen Sie eine neue Organisationseinheit unter \\\"\",[\"parentName\"],\"\\\" an.\"],\"vLMns2\":[\"Fehler:\"],\"vLyv1R\":[\"Ausblenden\"],\"vXIe7J\":[\"Sprache\"],\"vXhD8M\":[\"Mitarbeiter laden...\"],\"vmmz7E\":[\"Server-Version\"],\"w66diH\":[\"Teile \\\"\",[\"secretTitle\"],\"\\\"\"],\"wR67is\":[\"Ausstehende Synchronisierung\"],\"wy8uAs\":[\"Entwurf speichern\"],\"xBBUom\":[\"Sind Sie sicher, dass Sie den Zugang für \",[\"name\"],\" widerrufen wollen?\"],\"xCJdfg\":[\"Löschen\"],\"xGRPU0\":[\"<0>Hash-Kette:</0> Überprüft die sequentielle Integrität von Aktivitätsprotokollen. \\\"Ausstehend\\\" zeigt an, dass die Hash-Kette noch aufgebaut wird.<1/><2>Merkle Tree:</2> Batch-Verifizierung für einen effizienten Nachweis der Log-Einbindung (Sicherheitsstufen 2-3). Läuft in der Entwicklung jede Minute, in der Produktion stündlich.<3/><4>OpenTimestamp:</4> Bitcoin-Blockchain-Verankerung für unveränderlichen Existenznachweis (nur Sicherheitsstufe 3). Bitcoin-Bestätigung dauert ~10 Minuten.\"],\"y9c5t9\":[\"Optionale Beschreibung für diese Einheit.\"],\"yBgddv\":[\"Dokumente laden...\"],\"yHzSwP\":[\"← Zurück zu Mitarbeitern\"],\"yLKe22\":[\"Bis heute\"],\"yTjaVJ\":[\"Details zum Aktivitätsprotokoll\"],\"ysdDIA\":[\"Sie sind offline\"],\"yz7wBu\":[\"Schließen\"],\"z78wQd\":[\"Dateien hierher ziehen oder zum Durchsuchen anklicken\"],\"zH+1XO\":[\"Das Löschen von Organisationseinheiten ist im Offline-Modus nicht möglich. Bitte stellen Sie die Verbindung wieder her, um Änderungen vorzunehmen.\"],\"zVyXw4\":[\"Datei zu groß: \",[\"0\"],\". Die maximale Größe beträgt 10MB.\"],\"zga9sT\":[\"OK\"],\"zjo5MA\":[\"Organisationseinheit auswählen...\"],\"zmwvG2\":[\"Telefon\"],\"zzTOEt\":[\"Widerrufen...\"]}");
