+++ src/pages/ActivityLog/ActivityLogList.tsx
@@ -29,6 +29,7 @@ import {
 import { Badge } from "../../components/badge";
 import { Button } from "../../components/button";
 import { ActivityDetailDialog } from "./ActivityDetailDialog";
+import { VerificationDots } from "../../components/VerificationDots";
 
 /**
  * Security level badge component
@@ -47,83 +48,6 @@ function SecurityLevelBadge({ level }: { level: 1 | 2 | 3 }) {
   };
 
   return <Badge color={colors[level]}>{labels[level]}</Badge>;
-}
-
-/**
- * Verification status dots
- * Shows color-coded dots based on security level:
- * - Level 1: Hash Chain only (2 dots: data integrity + link integrity)
- * - Level 2: Hash Chain + Merkle Tree
- * - Level 3: Hash Chain + Merkle Tree + OpenTimestamp
- */
-function VerificationDots({ activity }: { activity: Activity }) {
-  const { verification, security_level } = activity;
-  const { _ } = useLingui();
-
-  // Helper to render a dot
-  const renderDot = (
-    status: boolean | null | undefined,
-    label: string,
-    notApplicable = false
-  ) => {
-    // Not applicable for this security level -> grey
-    if (notApplicable) {
-      return (
-        <span
-          className="inline-block w-2 h-2 rounded-full bg-zinc-300 dark:bg-zinc-600"
-          title={_(msg`${label}: N/A`)}
-        />
-      );
-    }
-
-    // Valid -> green
-    if (status === true) {
-      return (
-        <span
-          className="inline-block w-2 h-2 rounded-full bg-lime-500"
-          title={_(msg`${label}: Valid`)}
-        />
-      );
-    }
-
-    // Invalid -> red
-    if (status === false) {
-      return (
-        <span
-          className="inline-block w-2 h-2 rounded-full bg-red-500"
-          title={_(msg`${label}: Invalid`)}
-        />
-      );
-    }
-
-    // null or undefined = pending -> yellow
-    return (
-      <span
-        className="inline-block w-2 h-2 rounded-full bg-yellow-500"
-        title={_(msg`${label}: Pending`)}
-      />
-    );
-  };
-
-  return (
-    <div className="flex gap-1">
-      {/* Hash Chain - Data Integrity (always shown for all levels) */}
-      {renderDot(verification?.chain_valid, _(msg`Hash Chain (Data)`))}
-
-      {/* Hash Chain - Link Integrity (connection to predecessor) */}
-      {renderDot(verification?.chain_link_valid, _(msg`Hash Chain (Link)`))}
-
-      {/* Merkle Tree - show if Level 2+, or if data exists */}
-      {security_level >= 2 || verification?.merkle_valid !== undefined
-        ? renderDot(verification?.merkle_valid, _(msg`Merkle Tree`))
-        : renderDot(undefined, _(msg`Merkle Tree`), true)}
-
-      {/* OpenTimestamp - show if Level 3 only */}
-      {security_level >= 3
-        ? renderDot(verification?.ots_valid, _(msg`OpenTimestamp`))
-        : renderDot(undefined, _(msg`OpenTimestamp`), true)}
-    </div>
-  );
 }
 
 /**
@@ -513,7 +437,7 @@ export function ActivityLogList() {
                     <SecurityLevelBadge level={activity.security_level} />
                   </TableCell>
                   <TableCell className="w-16 shrink-0">
-                    <VerificationDots activity={activity} />
+                    <VerificationDots activity={activity} size="sm" />
                   </TableCell>
                 </TableRow>
               ))}
